move linear code from modules to functional
